test: rearrange the layout of tests folder, separate test concerns
