+++ tests/basetest.rs
@@ -1,53 +1,58 @@
-use cosign::configuration::{get_configuration, DatabaseSettings};
-use cosign::start_up::run;
-use uuid::Uuid;
-use sqlx::{Connection, Executor, PgConnection, PgPool};
-use std::{ net::TcpListener};
+/// basetest module containing functions to spawn a new instance of
+/// the application to facilitate ease in testing
 
-pub struct TestApplication {
-    pub address: String,
-    pub db_pool: PgPool,
-}
+pub mod base {
+    use cosign::configuration::{get_configuration, DatabaseSettings};
+    use cosign::start_up::run;
+    use uuid::Uuid;
+    use sqlx::{Connection, Executor, PgConnection, PgPool};
+    use std::{ net::TcpListener};
 
-/// Spawn an instance of the application
-pub async fn spawn_app() -> TestApplication {
-    let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
-    let port = listener.local_addr().unwrap().port();
-    let address = format!("http://127.0.0.1:{}", port);
-
-    let mut configuration = get_configuration().expect("Failed to load configuration");
-    configuration.database.database_name = Uuid::new_v4().to_string();
-    let connection_pool = configure_database(&configuration.database).await;
-
-    let server = run(listener, connection_pool.clone()).expect("Failed to bind address");
-    let _ = tokio::spawn(server);
-
-    TestApplication {
-        address,
-        db_pool: connection_pool,
+    pub struct TestApplication {
+        pub address: String,
+        pub db_pool: PgPool,
     }
-}
 
-pub async fn configure_database(config: &DatabaseSettings) -> PgPool {
-    // 1. Create database
-    let mut connection = PgConnection::connect(&config.connection_string_without_db())
-        .await
-        .expect("Failed to connect to Postgres.");
+    /// Spawn an instance of the application
+    pub async fn spawn_app() -> TestApplication {
+        let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
+        let port = listener.local_addr().unwrap().port();
+        let address = format!("http://127.0.0.1:{}", port);
 
-    connection
-        .execute(format!(r#"CREATE DATABASE "{}";"#, config.database_name).as_str())
-        .await
-        .expect("Failed to create database.");
+        let mut configuration = get_configuration().expect("Failed to load configuration");
+        configuration.database.database_name = Uuid::new_v4().to_string();
+        let connection_pool = configure_database(&configuration.database).await;
 
-    // 2. Migrate database
-    let connection_pool = PgPool::connect(&config.connection_string())
-        .await
-        .expect("Failed to connect to Postgres.");
+        let server = run(listener, connection_pool.clone()).expect("Failed to bind address");
+        let _ = tokio::spawn(server);
 
-    sqlx::migrate!("./migrations")
-        .run(&connection_pool)
-        .await
-        .expect("Failed to migrate the database");
+        TestApplication {
+            address,
+            db_pool: connection_pool,
+        }
+    }
 
-    connection_pool
+    pub async fn configure_database(config: &DatabaseSettings) -> PgPool {
+        // 1. Create database
+        let mut connection = PgConnection::connect(&config.connection_string_without_db())
+            .await
+            .expect("Failed to connect to Postgres.");
+
+        connection
+            .execute(format!(r#"CREATE DATABASE "{}";"#, config.database_name).as_str())
+            .await
+            .expect("Failed to create database.");
+
+        // 2. Migrate database
+        let connection_pool = PgPool::connect(&config.connection_string())
+            .await
+            .expect("Failed to connect to Postgres.");
+
+        sqlx::migrate!("./migrations")
+            .run(&connection_pool)
+            .await
+            .expect("Failed to migrate the database");
+
+        connection_pool
+    }
 }
