+++ tests/collect_xpubs_test.rs
@@ -1,109 +1,13 @@
 pub mod basetest;
 
-use basetest::spawn_app;
-use reqwest;
-use std::{collections::HashMap};
+use basetest::base;
+use std::collections::HashMap;
 
-
-/// test the ping endpoint to confirm the server is running
-#[tokio::test]
-async fn ping_test() {
-    // 1. Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-
-    // 2. Act
-    let response = client
-        .get(format!("{}/ping", &test_app.address))
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    // 3. Assert
-    assert!(response.status().is_success());
-    assert_eq!(Some(0), response.content_length());
-}
-
-/// Test user account creation
-#[tokio::test]
-async fn create_user_returns_201_valid_json_data_test() {
-    // 1. Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-
-    // 2. Act
-    let mut body = HashMap::new();
-    body.insert("email".to_string(), "user@email.com".to_string());
-    body.insert("password".to_string(), "password".to_string());
-
-    let url = format!("{}/create_user", &test_app.address);
-
-    let response = client
-        .post(&url)
-        .json(&body)
-        .send()
-        .await
-        .expect("Failed to execute request");
-
-    // 3. Assert
-    assert_eq!(201, response.status().as_u16());
-
-    let saved = sqlx::query!("SELECT email FROM users",)
-        .fetch_one(&test_app.db_pool)
-        .await
-        .expect("Failed to fetch saved user");
-
-    assert_eq!(saved.email, "user@email.com")
-}
-
-/// Test user creation with bad input data
-#[tokio::test]
-async fn create_user_returns_400_json_fields_present_but_empty_test() {
-    // 1. Arrange
-    let test_app = spawn_app().await;
-    let client = reqwest::Client::new();
-    let test_cases = vec![
-        (
-            HashMap::from([
-                ("email".to_string(), "".to_string()),
-                ("password".to_string(), "secret".to_string()),
-            ]),
-            "email cannot be empty",
-        ),
-        (
-            HashMap::from([
-                ("email".to_string(), "janedoe@email.com".to_string()),
-                ("password".to_string(), "".to_string()),
-            ]),
-            "password cannot be empty",
-        ),
-    ];
-
-    let url = format!("{}/create_user", &test_app.address);
-    for (invalid_body, error_msg) in test_cases {
-        // 2. Act
-        let response = client
-            .post(&url)
-            .json(&invalid_body)
-            .send()
-            .await
-            .expect("Failed to execute request");
-
-        // 3. Assert
-        assert_eq!(
-            400,
-            response.status().as_u16(),
-            "The API failed with 400 Bad Request because {}.",
-            error_msg
-        );
-    }
-}
-
-/// Test that existing users can uoload two xpubs
+/// Test that existing users can upload two xpubs
 #[tokio::test]
 async fn collect_xpubs_returns_200_for_existing_user() {
     // 1. Arrange
-    let test_app = spawn_app().await;
+    let test_app = base::spawn_app().await;
     let client = reqwest::Client::new();
     let create_user_url = format!("{}/create_user", &test_app.address);
     let collect_xpub_url = format!("{}/collect_xpubs", &test_app.address);
@@ -160,7 +64,7 @@ async fn collect_xpubs_returns_200_for_existing_user() {
 #[tokio::test]
 async fn collect_xpubs_returns_403_for_nonexistent_users() {
     // 1. Arrange
-    let test_app = spawn_app().await;
+    let test_app = base::spawn_app().await;
     let client = reqwest::Client::new();
     let create_user_url = format!("{}/create_user", &test_app.address);
     let collect_xpub_url = format!("{}/collect_xpubs", &test_app.address);
@@ -210,7 +114,7 @@ async fn collect_xpubs_returns_403_for_nonexistent_users() {
 #[tokio::test]
 async fn collect_xpubs_returns_400_for_invalid_xpub() {
     // 1. Arrange
-    let test_app = spawn_app().await;
+    let test_app = base::spawn_app().await;
     let client = reqwest::Client::new();
     let collect_xpub_url = format!("{}/collect_xpubs", &test_app.address);
     let mut xpub_body = HashMap::new();
